all classes back into one module, some additional renamings
