--- kicadfield.py
@@ -1,16 +1,0 @@
-
-
-
-
-
-class KicadField(object):
-	def __init__(self):
-		self.Aliases = []
-		self.name = ""
-		self.contents = ""
-
-	def appendAlias(self,newAlias):
-		self.Aliases.append(newAlias)
-
-
-
